Heavy Cleanup and Refactoring
To prepare for the pull request, I decided to do some heavy cleanup and
refactoring.

* Removed all commented out code in isoTile.js and main.js
* JSLinted the file isoTile.js  and main.js
* Added descriptive comments to isoTile.js
* Added JSDoc documentation.
* Added global declarations for JSLint to the top of the files.
* Put everything related to the engine in a namespace object called SEI
(not finished).
* Refactored a lot of variable names to be more descriptive.
* In the SetChildren function in isoTile.js, I changed async.concat to
_.flatten as I have seen previously that using async.concat to flatten
arrays is slower than underscores.
* Made it so instead of outputting the fps to console, that it would
instead change a p tag above the canvas to the last seconds frame
count.  This is because I found out through testing that, over time, as
more log are in the console, the more it effects performance.